vimrc: switch statusline to display on top
The intent of this change is to avoid issues where when the statusline
is on the bottom, it can be clipped and omit the first part of the path
or other info when using 2/3+ vertical panes. By putting it on the top
it ensures that the active pane has full information always.