vim: switch from deoplete to coc for completion
This is motivated by issues dealing with the python toolchain that
deoplete depends on that are too much of a pain to work through in
various environments and machines with lots of issues with incompatible
pip installs with different python versions and vim linkages through
brew, etc.

Trying this out seems to "just work" in many ways and is a lot simpler,
shedding the more complex layer of python neovim emulation.

First impressions are that the resulting UX is quite snappy as well with
some interesting changes that may well be considered improvements.