home/vim: replace ale, git-gutter, vim-flow with coc.nvim config
This further cements focus around the coc.nvim tool to reduce the number
of packages installed through vim-plug and the number of tools running
in parallel to focus specifically on coc.nvim as the way in which
linting, completion, etc. happens within the editor.

The next commit will extend this further for Go code.