vimrc: remove 'for' directives for go plugins
This is no longer recommended in the vim-plug docs:
https://github.com/junegunn/vim-plug/wiki/faq#when-should-i-use-on-or-for-option

And having these always available lets them be used in other cases
besides just Go source files, like templates. Worth noting that an
alternative would be to specify 'go*' as it seems wildcards are
supported, but this approach seems cleaner overall.

It may make sense to do similar change to other uses of 'for' but that
will require looking at each plugin.